+++ src/VpnApiModule.php
@@ -36,33 +36,15 @@ class VpnApiModule implements ServiceModuleInterface
     /** @var \DateInterval */
     private $sessionExpiry;
 
-    /** @var bool */
-    private $pickRemotesAtRandom = true;
-
     /** @var \DateTime */
     private $dateTime;
 
-    /**
-     * @param \LC\Common\Config                  $config
-     * @param \LC\Common\HttpClient\ServerClient $serverClient
-     * @param \DateInterval                      $sessionExpiry
-     */
     public function __construct(Config $config, ServerClient $serverClient, DateInterval $sessionExpiry)
     {
         $this->config = $config;
         $this->serverClient = $serverClient;
         $this->sessionExpiry = $sessionExpiry;
         $this->dateTime = new DateTime();
-    }
-
-    /**
-     * @param bool $pickRemotesAtRandom
-     *
-     * @return void
-     */
-    public function setPickRemotesAtRandom($pickRemotesAtRandom)
-    {
-        $this->pickRemotesAtRandom = $pickRemotesAtRandom;
     }
 
     /**
@@ -190,6 +172,7 @@ class VpnApiModule implements ServiceModuleInterface
                 $accessTokenInfo = $hookData['auth'];
                 try {
                     $requestedProfileId = InputValidation::profileId($request->getQueryParameter('profile_id'));
+                    $pickMode = ClientConfig::validatePickMode((int) $request->getQueryParameter('pick_mode', false, ClientConfig::MODE_RANDOM));
                     $profileList = $this->serverClient->getRequireArray('profile_list');
                     $userPermissions = $this->getPermissionList($accessTokenInfo);
 
@@ -210,7 +193,7 @@ class VpnApiModule implements ServiceModuleInterface
                         return new ApiErrorResponse('profile_config', 'user has no access to this profile');
                     }
 
-                    return $this->getConfigOnly($requestedProfileId);
+                    return $this->getConfigOnly($requestedProfileId, $pickMode);
                 } catch (InputValidationException $e) {
                     return new ApiErrorResponse('profile_config', $e->getMessage());
                 }
@@ -263,10 +246,11 @@ class VpnApiModule implements ServiceModuleInterface
 
     /**
      * @param string $profileId
+     * @param int    $pickMode
      *
      * @return Response
      */
-    private function getConfigOnly($profileId)
+    private function getConfigOnly($profileId, $pickMode)
     {
         // obtain information about this profile to be able to construct
         // a client configuration file
@@ -276,7 +260,7 @@ class VpnApiModule implements ServiceModuleInterface
         // get the CA & tls-auth
         $serverInfo = $this->serverClient->getRequireArray('server_info');
 
-        $clientConfig = ClientConfig::get($profileData, $serverInfo, [], $this->pickRemotesAtRandom);
+        $clientConfig = ClientConfig::get($profileData, $serverInfo, [], $pickMode);
         $clientConfig = str_replace("\n", "\r\n", $clientConfig);
 
         $response = new Response(200, 'application/x-openvpn-profile');
@@ -286,8 +270,6 @@ class VpnApiModule implements ServiceModuleInterface
     }
 
     /**
-     * @param \LC\Portal\OAuth\VpnAccessTokenInfo $accessTokenInfo
-     *
      * @return array
      */
     private function getCertificate(VpnAccessTokenInfo $accessTokenInfo)
@@ -345,8 +327,6 @@ class VpnApiModule implements ServiceModuleInterface
     }
 
     /**
-     * @param \LC\Portal\OAuth\VpnAccessTokenInfo $accessTokenInfo
-     *
      * @return array<string>
      */
     private function getPermissionList(VpnAccessTokenInfo $accessTokenInfo)
@@ -359,8 +339,6 @@ class VpnApiModule implements ServiceModuleInterface
     }
 
     /**
-     * @param \LC\Portal\OAuth\VpnAccessTokenInfo $accessTokenInfo
-     *
      * @return \DateTime
      */
     private function getExpiresAt(VpnAccessTokenInfo $accessTokenInfo)