+++ src/VpnPortalModule.php
@@ -45,9 +45,6 @@ class VpnPortalModule implements ServiceModuleInterface
     /** @var \fkooman\OAuth\Server\ClientDbInterface */
     private $clientDb;
 
-    /** @var bool */
-    private $pickRemotesAtRandom = true;
-
     /** @var \DateTime */
     private $dateTime;
 
@@ -68,16 +65,6 @@ class VpnPortalModule implements ServiceModuleInterface
     public function setDateTime(DateTime $dateTime)
     {
         $this->dateTime = $dateTime;
-    }
-
-    /**
-     * @param bool $pickRemotesAtRandom
-     *
-     * @return void
-     */
-    public function setPickRemotesAtRandom($pickRemotesAtRandom)
-    {
-        $this->pickRemotesAtRandom = $pickRemotesAtRandom;
     }
 
     /**
@@ -391,7 +378,7 @@ class VpnPortalModule implements ServiceModuleInterface
         // get the CA & tls-auth
         $serverInfo = $this->serverClient->getRequireArray('server_info');
 
-        $clientConfig = ClientConfig::get($profileData, $serverInfo, $clientCertificate, $this->pickRemotesAtRandom);
+        $clientConfig = ClientConfig::get($profileData, $serverInfo, $clientCertificate, ClientConfig::MODE_RANDOM);
 
         // convert the OpenVPN file to "Windows" format, no platform cares, but
         // in Notepad on Windows it looks not so great everything on one line